+++ catalog/model/payment/paymos.php
@@ -26,19 +26,4 @@ class Paymos extends \Opencart\System\Engine\Model
             'sort_order' => (int) $this->config->get('payment_paymos_sort_order'),
         ];
     }
-
-    public function getMethod(array $address = []): array
-    {
-        $methods = $this->getMethods($address);
-        if (!$methods) {
-            return [];
-        }
-
-        return [
-            'code' => $methods['code'],
-            'title' => $methods['name'],
-            'terms' => '',
-            'sort_order' => $methods['sort_order'],
-        ];
-    }
 }